refactor(utils): split shared helpers by domain
Move the broad utils module into domain-focused helpers for error coercion, file walking/hash, ingest filtering, KB path validation, and frontmatter parsing. Update source imports and architecture anchors to use the narrower module boundaries.

Closes #135
